+++ Storage/Disk/Track/TrackSerialiser.hpp
@@ -10,6 +10,8 @@
 #define TrackSerialiser_h
 
 #include "../DPLL/DigitalPhaseLockedLoop.hpp"
+#include "PCMSegment.hpp"
+#include "Track.hpp"
 
 namespace Storage {
 namespace Disk {
@@ -28,42 +30,7 @@ namespace Disk {
 	@param length_of_a_bit The expected length of a single bit, as a proportion of the
 	track length.
 */
-template <typename T> PCMSegment track_serialisation(T &track, Time length_of_a_bit) {
-	DigitalPhaseLockedLoop pll(100, 16);
-
-	struct ResultAccumulator: public DigitalPhaseLockedLoop::Delegate {
-		PCMSegment result;
-		void digital_phase_locked_loop_output_bit(int value) {
-			result.data.resize(1 + (result.number_of_bits >> 3));
-			if(value) result.data[result.number_of_bits >> 3] |= 0x80 >> (result.number_of_bits & 7);
-			result.number_of_bits++;
-		}
-	} result_accumulator;
-	pll.set_delegate(&result_accumulator);
-	result_accumulator.result.number_of_bits = 0;
-	result_accumulator.result.length_of_a_bit = length_of_a_bit;
-
-	Time length_multiplier = Time(100*length_of_a_bit.clock_rate, length_of_a_bit.length);
-	length_multiplier.simplify();
-
-	// start at the index hole
-	track.seek_to(Time(0));
-
-	// grab events until the next index hole
-	Time time_error = Time(0);
-	while(true) {
-		Track::Event next_event = track.get_next_event();
-		if(next_event.type == Track::Event::IndexHole) break;
-
-		Time extended_length = next_event.length * length_multiplier + time_error;
-		time_error.clock_rate = extended_length.clock_rate;
-		time_error.length = extended_length.length % extended_length.clock_rate;
-		pll.run_for(Cycles((int)extended_length.get_unsigned_int()));
-		pll.add_pulse();
-	}
-
-	return result_accumulator.result;
-}
+PCMSegment track_serialisation(Track &track, Time length_of_a_bit);
 
 }
 }
